+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 2024-07-31
 
+### Added
+
+- The ability to (also) remember Bible verses.
+
 ### Changed
 
 - Cleaned up some code.